Protecting react routes

+++ frontend/src/components/WorkoutDetails.js
@@ -1,12 +1,20 @@
 import React from "react";
 import {useWorkoutsContext} from "../hooks/useWorkoutsContext"
 import formatDistanceToNow from "date-fns/formatDistanceToNow"
+import { useAuthContext } from "../hooks/useAuthContext";
 
 const WorkoutDetails = (props) => {
   const {dispatch} = useWorkoutsContext()
+  const {user} = useAuthContext()
   const handleClick = async ()=>{
+    if(!user){
+      return
+    }
     const response = await fetch("/api/workouts/" + props.workout._id, {
-      method:"DELETE"
+      method:"DELETE",
+      headers:{
+        "Authorization":`Bearer ${user.token}`
+      }
     })
     const json = await response.json()
     if(response.ok){
